create a guest user
setting the guest attribute to true means that it will be allowed to be
saved, even without an email address. it won't be possible to update
this user without adding an email address, though. if you need to
update guest users, you will want to add a guest column to the users
table so that the "guest" status will be saved.